+++ app/src/settings_view/platform/expire_api_key_button.rs
@@ -4,7 +4,6 @@ use warpui::ui_components::components::UiComponent;
 use warpui::{AppContext, Element, Entity, SingletonEntity, TypedActionView, View, ViewContext};
 
 use crate::server::ids::ApiKeyUid;
-use crate::server::server_api::auth::AuthClient;
 use crate::ui_components::buttons::icon_button;
 use crate::ui_components::icons::Icon;
 
@@ -46,10 +45,11 @@ impl ExpireApiKeyButton {
         self.request_state = RequestState::Pending;
         ctx.notify();
 
-        let server_api = crate::server::server_api::ServerApiProvider::as_ref(ctx).get();
+        let auth_client =
+            crate::server::server_api::ServerApiProvider::as_ref(ctx).get_auth_client();
         let uid_for_req = self.key_uid.clone();
         ctx.spawn(
-            async move { server_api.expire_api_key(&uid_for_req).await },
+            async move { auth_client.expire_api_key(&uid_for_req).await },
             move |me, res, ctx| match res {
                 Ok(
                     warp_graphql::mutations::expire_api_key::ExpireApiKeyResult::ExpireApiKeyOutput(
