+++ crates/warp_server_client/src/base_client.rs
@@ -16,40 +16,95 @@ use crate::auth::{AgentIdentity, UserAuthenticationError};
 #[cfg_attr(not(target_family = "wasm"), async_trait)]
 #[cfg_attr(target_family = "wasm", async_trait(?Send))]
 pub trait BaseClient: Send + Sync {
+    /// Returns the HTTP transport used to send server API requests.
+    ///
+    /// Extracted client implementations should use this client rather than
+    /// constructing their own transport so application-level HTTP setup remains shared.
     fn http_client(&self) -> Arc<http_client::Client>;
 
+    /// Returns the anonymous installation identifier used to correlate unauthenticated requests.
+    ///
+    /// Endpoint implementations should add this identifier to requests whose
+    /// protocol includes anonymous experiment or pre-login identity handling.
     fn anonymous_id(&self) -> String;
 
+    /// Returns GraphQL request options for a request that does not use the logged-in credentials.
+    ///
+    /// Clients may extend these options with request-specific headers or tokens, such
+    /// as the explicit token supplied while fetching a newly authenticated user.
     fn unauthenticated_graphql_request_options(&self) -> RequestOptions;
 
+    /// Returns GraphQL request options for an authenticated operation.
+    ///
+    /// Extracted GraphQL clients should use this method through the shared request
+    /// helper so token refresh, timeouts, and application-owned headers remain centralized.
     async fn graphql_request_options(&self, timeout: Option<Duration>) -> Result<RequestOptions>;
 
+    /// Exchanges login credentials for the credential state used by authenticated requests.
+    ///
+    /// Auth client implementations invoke this while completing login; the application
+    /// remains responsible for Firebase, session, and other credential lifecycle behavior.
     async fn exchange_credentials(
         &self,
         token: LoginToken,
     ) -> StdResult<Credentials, UserAuthenticationError>;
 
+    /// Returns a currently valid authentication token, refreshing it when required.
+    ///
+    /// Extracted clients should use this for non-GraphQL authenticated requests or APIs
+    /// that expose token retrieval as part of their public client contract.
     async fn get_or_refresh_access_token(&self) -> Result<AuthToken>;
 
+    /// Starts the OAuth device authorization flow and returns its server-issued device code.
+    ///
+    /// This capability remains application-provided because the OAuth client and its
+    /// platform-specific execution details are not owned by extracted API crates.
     async fn request_device_code(
         &self,
     ) -> StdResult<oauth2::StandardDeviceAuthorizationResponse, UserAuthenticationError>;
 
+    /// Completes an OAuth device authorization flow and returns a Firebase login token.
+    ///
+    /// Auth client implementations expose this operation while the application supplies
+    /// the configured OAuth transport and timer behavior.
     async fn exchange_device_access_token(
         &self,
         details: &oauth2::StandardDeviceAuthorizationResponse,
         timeout: Duration,
     ) -> StdResult<FirebaseToken, UserAuthenticationError>;
 
+    /// Lists public agent identities available to API-key creation flows.
+    ///
+    /// This is a base capability until its public REST endpoint is extracted alongside
+    /// the GraphQL-backed API client methods that consume its result.
     async fn list_agent_identities(&self) -> Result<Vec<AgentIdentity>>;
 
+    /// Returns an ambient workload token when the current runtime supports issuing one.
+    ///
+    /// Extracted clients surface this for ambient-agent authentication while leaving
+    /// workload-token caching and platform integration in the application.
     async fn get_or_create_ambient_workload_token(&self) -> Result<Option<String>>;
 
+    /// Returns whether authentication failures may be handled as refreshable user-session failures.
+    ///
+    /// The shared GraphQL request helper uses this distinction to avoid emitting
+    /// user-session events for externally managed credentials.
     fn is_auth_refresh_allowed(&self) -> bool;
 
+    /// Notifies the application that a GraphQL request was blocked by staging access controls.
+    ///
+    /// The shared request helper calls this hook instead of depending on application event types.
     fn on_graphql_staging_access_blocked(&self);
 
+    /// Notifies the application that a GraphQL request received an IAP challenge.
+    ///
+    /// The shared request helper calls this hook so the application can refresh
+    /// its IAP state without exposing that state to extracted clients.
     fn on_graphql_iap_challenge_received(&self);
 
+    /// Notifies the application that a GraphQL response indicates a disabled user account.
+    ///
+    /// The shared request helper only invokes this for refreshable user sessions;
+    /// callers using externally managed credentials receive an authentication error instead.
     fn on_graphql_user_account_disabled(&self);
 }
